Add styling for the new game buttons
I added a little bit more style to the New Game buttons, as well as making sure that the button text color stayed black when clicked, because on a Mac, I have noticed that the color would change to white when the button was clicked. This should fix that issue.